added hire date

+++ src/main/java/com/employees/repositories/DbInit.java
@@ -5,7 +5,6 @@ import com.employees.models.AccountInformation;
 import com.employees.models.Degree;
 import com.employees.models.Employee;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -13,6 +12,8 @@ import org.springframework.stereotype.Service;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.Iterator;
 
 @Service
 @AllArgsConstructor
@@ -23,8 +24,16 @@ public class DbInit implements CommandLineRunner {
 
     @Override
     public void run(String... args) throws InternalException, ParseException {
-        accountInformationRepository.deleteAll();
-        employeeRepository.deleteAll();
+        Iterable<AccountInformation> account = accountInformationRepository.findAll();
+        Iterator<AccountInformation> accounts = account.iterator();
+        if(accounts.hasNext()) {
+            accountInformationRepository.deleteAll();
+        }
+        Iterable<Employee> emp = employeeRepository.findAll();
+        Iterator<Employee> employees = emp.iterator();
+        if(employees.hasNext()) {
+            employeeRepository.deleteAll();
+        }
         Employee admin = new Employee();
         admin.setFirstName("admin");
         admin.setGrossSalary(13000.0);
@@ -34,6 +43,9 @@ public class DbInit implements CommandLineRunner {
         admin.setDegree(Degree.SENIOR);
         admin.setNationalId(1000L);
         admin.setDob(new SimpleDateFormat("dd/MM/yyyy").parse("1/1/1980"));
+        long millis=System.currentTimeMillis();
+        Date now = new Date(millis);
+        admin.setHireDate(now);
         employeeRepository.save(admin);
         AccountInformation accountInformation = new AccountInformation();
         accountInformation.setEmployee(admin);
@@ -55,6 +67,7 @@ public class DbInit implements CommandLineRunner {
         employee.setDegree(Degree.FRESH);
         employee.setNationalId(1009L);
         employee.setDob(new SimpleDateFormat("dd/MM/yyyy").parse("1/1/1988"));
+        employee.setHireDate(now);
         employeeRepository.save(employee);
         AccountInformation accountInformation1 = new AccountInformation();
         accountInformation1.setEmployee(employee);
